requirements.txt
Flask==3.1.0
google-cloud-storage==2.19.0
tensorflow==2.18.0
numpy==2.0.2
pillow==11.0.0
google-generativeai==0.8.3